feat: add new Justiciar abilities and change mastery variable type

+++ src/main/java/me/vangoo/domain/abilities/core/IAbilityContext.java
@@ -142,7 +142,7 @@ public interface IAbilityContext {
 
     int getExperienceLevel(UUID targetId);
 
-    int getBeyonderMastery(UUID targetId);
+    double getBeyonderMastery(UUID targetId);
 
     // ==========================================
     // VISUAL EFFECTS (EffectLib)
